consent-ledger/pyproject.toml
[tool.poetry]
name = "consent-ledger"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["Gokul G <gane.incore@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^4.0.0"
python-dotenv = "^1.0.0"
algorand-python = "^3"
algorand-python-testing = "^1"

[tool.poetry.group.dev.dependencies]
algokit-client-generator = "^2.1.0"
puyapy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

